requirements.txt
python-telegram-bot==21.3
anthropic>=0.30.0
python-dotenv==1.0.1

# Google Docs support (optional but recommended)
google-api-python-client==2.136.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
